+++ CHANGELOG.md
@@ -22,6 +22,11 @@ must be regular text
 * Add `ResendableContent#createResends` which create adapted list of resends for content
 * Add `TextContent` own `createResends` realisation
 
+### 0.9.2
+
+* `RequestsExecutor#executeAsync(Request, CoroutineScope)` now will return `Deferred` for cases when you need result
+* `RequestsExecutor#executeUnsafe` will automatically retry request if it was unsuccessful and retries > 0
+
 ### 0.8.5
 
 * Add extension `String#toMarkdown`